+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 ### Added
 - Refactor security tests.
+- Add placeholder controllers for WebDAV HTTP methods.
 
 ## 0.0.5 - 2023-05-13
 ### Added